Introduce CameraNeoService and session controllers; remove CameraNeo monolith
Manifest points foreground service at CameraNeoService.
